Rename queryBuilder to query-builder
