+++ src/platforms/salesforce/actions/query-builder.ts
